tests/wpt/tests/css/css-inline/parsing/vertical-align-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Inline Layout: getComputedStyle().verticalAlign</title>
<link rel="help" href="https://www.w3.org/TR/CSS2/visudet.html#propdef-vertical-align">
<link rel="help" href="https://drafts.csswg.org/css-inline/#propdef-vertical-align">
<meta name="assert" content="vertical-align computed value is as specified, with lengths made absolute.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
// Test shorthand values equivalent to the initial `baseline` value
test_computed_value("vertical-align", "baseline", "baseline");
test_computed_value("vertical-align", "0", "baseline");
test_computed_value("vertical-align", "0px", "baseline");
test_computed_value("vertical-align", "baseline 0", "baseline");
test_computed_value("vertical-align", "0px baseline", "baseline");

// Test individual non-initial values for `baseline-source`
test_computed_value("vertical-align", "first");
test_computed_value("vertical-align", "last");

// Test individual non-initial values for `alignment-baseline`
test_computed_value("vertical-align", "text-bottom");
test_computed_value("vertical-align", "alphabetic");
test_computed_value("vertical-align", "ideographic");
test_computed_value("vertical-align", "middle");
test_computed_value("vertical-align", "central");
test_computed_value("vertical-align", "mathematical");
test_computed_value("vertical-align", "text-top");

// Test individual non-initial values for `baseline-shift`
test_computed_value("vertical-align", "-10px");
test_computed_value("vertical-align", "20%");
test_computed_value("vertical-align", "calc(20% + 10px)");
test_computed_value("vertical-align", "calc(10px - 0.5em)", "-10px");
test_computed_value("vertical-align", "sub");
test_computed_value("vertical-align", "super");
test_computed_value("vertical-align", "top");
test_computed_value("vertical-align", "center");
test_computed_value("vertical-align", "bottom");
</script>
</body>
</html>
